requirements.txt
flask
pandas
requests
rdkit
pillow
umap-learn
plotly

